Add files via upload
This is an agent which answers users' queries in an insightful Harry Potter context. This makes learning fun and interesting. We have used Serper AI for RAG and Gemini API. A persona of J.K Rowling was created and defining its role and goal. Temperature was set to 0.6 for creativity. To reduce hallucination, i have set Verbose = True, for Reasoning + Acting (ReAct).